+++ sporktris.h
@@ -7,16 +7,16 @@
 
 struct Tetromino {
   struct Rect {
-    byte x, y, x2, y2;
-    byte center_x();
+    uint8_t x, y, x2, y2;
+    uint8_t center_x();
   };
   
-  byte positions;
-  byte axis_x;
-  byte axis_y;
-  byte cur_pos;
-  byte color;
-  byte points[4][2];
+  uint8_t positions;
+  uint8_t axis_x;
+  uint8_t axis_y;
+  uint8_t cur_pos;
+  uint8_t color;
+  uint8_t points[4][2];
 
   Tetromino rotated(bool cw);
   Rect bounding_rect();
@@ -33,11 +33,11 @@ class Sporktris: public InputProcessor {
   bool play();
 
   private:
-  byte *board;
+  uint8_t *board;
   
   Tetromino cur_piece;
-  short piece_x;
-  short piece_y;
+  int8_t piece_x;
+  int8_t piece_y;
   bool need_new_piece;
   bool clearing_lines;
